+++ common/src/main/java/de/cech12/usefulhats/item/BunnyEarsItem.java
@@ -29,13 +29,17 @@ public class BunnyEarsItem extends AbstractHatItem implements IEquipmentChangeLi
     public void appendHoverText(@NotNull ItemStack stack, @NotNull Item.TooltipContext context, @NotNull List<Component> tooltip, @NotNull TooltipFlag flagIn) {
         super.appendHoverText(stack, context, tooltip, flagIn);
         int enchantmentLevel = Services.PLATFORM.getEnchantmentLevel(stack, Enchantments.EFFICIENCY) + 1;
-        tooltip.add(Component.translatable("item.usefulhats.bunny_ears.desc.jump_boost", UsefulHatsUtils.getRomanNumber(enchantmentLevel, false)).withStyle(ChatFormatting.BLUE));
-        tooltip.add(Component.translatable("item.usefulhats.bunny_ears.desc.eating", enchantmentLevel + 1).withStyle(ChatFormatting.BLUE));
+        if (Services.CONFIG.isBunnyEarsJumpBoostEnabled()) {
+            tooltip.add(Component.translatable("item.usefulhats.bunny_ears.desc.jump_boost", UsefulHatsUtils.getRomanNumber(enchantmentLevel, false)).withStyle(ChatFormatting.BLUE));
+        }
+        if (Services.CONFIG.isBunnyEarsEatBoostEnabled()) {
+            tooltip.add(Component.translatable("item.usefulhats.bunny_ears.desc.eating", enchantmentLevel + 1).withStyle(ChatFormatting.BLUE));
+        }
     }
 
     @Override
-    public void inventoryTick(@NotNull ItemStack stack, Level level, @NotNull Entity entity, int slot, boolean selectedIndex) {
-        if (!level.isClientSide && entity instanceof LivingEntity livingEntity) {
+    public void inventoryTick(@NotNull ItemStack stack, @NotNull Level level, @NotNull Entity entity, int slot, boolean selectedIndex) {
+        if (Services.CONFIG.isBunnyEarsJumpBoostEnabled() && !level.isClientSide && entity instanceof LivingEntity livingEntity) {
             if (!Services.REGISTRY.getEquippedHatItemStacks(livingEntity).contains(stack)) return; //only one worn stack of this item should add its effect
             int amplifier = Services.PLATFORM.getEnchantmentLevel(stack, Enchantments.EFFICIENCY);
             if (!this.isEffectCausedByOtherSource(livingEntity, MobEffects.JUMP, JUMP_BOOST_DURATION, amplifier)) {
@@ -48,7 +52,7 @@ public class BunnyEarsItem extends AbstractHatItem implements IEquipmentChangeLi
 
     @Override
     public int onItemUseEventStart(LivingEntity entity, ItemStack usedStack, int actualDuration, ItemStack headSlotItemStack) {
-        if (usedStack.has(DataComponents.FOOD)) {
+        if (Services.CONFIG.isBunnyEarsEatBoostEnabled() && usedStack.has(DataComponents.FOOD)) {
             int amplifier = Services.PLATFORM.getEnchantmentLevel(headSlotItemStack, Enchantments.EFFICIENCY) + 2;
             this.damageHatItemByOne(headSlotItemStack, entity);
             return actualDuration / amplifier;
@@ -58,16 +62,20 @@ public class BunnyEarsItem extends AbstractHatItem implements IEquipmentChangeLi
 
     @Override
     public void onUnequippedHatItem(LivingEntity entity, ItemStack oldStack) {
-        // disable effects when hat is removed from slot
-        int amplifier = Services.PLATFORM.getEnchantmentLevel(oldStack, Enchantments.EFFICIENCY);
-        this.removeEffect(entity, MobEffects.JUMP, JUMP_BOOST_DURATION, amplifier);
+        if (Services.CONFIG.isBunnyEarsJumpBoostEnabled()) {
+            // disable effects when hat is removed from slot
+            int amplifier = Services.PLATFORM.getEnchantmentLevel(oldStack, Enchantments.EFFICIENCY);
+            this.removeEffect(entity, MobEffects.JUMP, JUMP_BOOST_DURATION, amplifier);
+        }
     }
 
     @Override
     public void onLivingJumpEvent(LivingEntity jumpingEntity, ItemStack headSlotItemStack) {
-        int amplifier = Services.PLATFORM.getEnchantmentLevel(headSlotItemStack, Enchantments.EFFICIENCY);
-        if (!this.isEffectCausedByOtherSource(jumpingEntity, MobEffects.JUMP, JUMP_BOOST_DURATION, amplifier)) {
-            this.damageHatItemByOne(headSlotItemStack, jumpingEntity);
+        if (Services.CONFIG.isBunnyEarsJumpBoostEnabled()) {
+            int amplifier = Services.PLATFORM.getEnchantmentLevel(headSlotItemStack, Enchantments.EFFICIENCY);
+            if (!this.isEffectCausedByOtherSource(jumpingEntity, MobEffects.JUMP, JUMP_BOOST_DURATION, amplifier)) {
+                this.damageHatItemByOne(headSlotItemStack, jumpingEntity);
+            }
         }
     }
 }