+++ fabric/src/main/java/de/cech12/usefulhats/platform/FabricConfigHelper.java
@@ -57,6 +57,10 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
         @ConfigEntry.Gui.Tooltip(count = 2)
         @ConfigEntry.BoundedDiscrete(min = BUNNY_EARS_DURABILITY_MIN, max = BUNNY_EARS_DURABILITY_MAX)
         public long DURABILITY = BUNNY_EARS_DURABILITY_DEFAULT;
+        @ConfigEntry.Gui.Tooltip(count = 3)
+        public boolean EAT_BOOST_ENABLED = BUNNY_EARS_EAT_BOOST_ENABLED_DEFAULT;
+        @ConfigEntry.Gui.Tooltip(count = 3)
+        public boolean JUMP_BOOST_ENABLED = BUNNY_EARS_JUMP_BOOST_ENABLED_DEFAULT;
 
         private BunnyEars() {}
     }
@@ -359,6 +363,16 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     @Override
     public int getBunnyEarsDurability() {
         return getInteger(getConfig().getConfig().BUNNY_EARS.DURABILITY);
+    }
+
+    @Override
+    public boolean isBunnyEarsEatBoostEnabled() {
+        return getConfig().BUNNY_EARS.EAT_BOOST_ENABLED;
+    }
+
+    @Override
+    public boolean isBunnyEarsJumpBoostEnabled() {
+        return getConfig().BUNNY_EARS.JUMP_BOOST_ENABLED;
     }
 
     @Override