+++ neoforge/src/main/java/de/cech12/usefulhats/platform/NeoForgeConfigHelper.java
@@ -30,6 +30,8 @@ public class NeoForgeConfigHelper implements IConfigHelper {
 
     public static final ModConfigSpec.BooleanValue BUNNY_EARS_DAMAGE_ENABLED;
     public static final ModConfigSpec.IntValue BUNNY_EARS_DURABILITY;
+    public static final ModConfigSpec.BooleanValue BUNNY_EARS_EAT_BOOST_ENABLED;
+    public static final ModConfigSpec.BooleanValue BUNNY_EARS_JUMP_BOOST_ENABLED;
 
     public static final ModConfigSpec.BooleanValue CHOPPING_HAT_DAMAGE_ENABLED;
     public static final ModConfigSpec.IntValue CHOPPING_HAT_DURABILITY;
@@ -128,6 +130,12 @@ public class NeoForgeConfigHelper implements IConfigHelper {
         BUNNY_EARS_DURABILITY = builder
                 .comment(BUNNY_EARS_DURABILITY_DESCRIPTION)
                 .defineInRange("bunnyEarsDurability", BUNNY_EARS_DURABILITY_DEFAULT, BUNNY_EARS_DURABILITY_MIN, BUNNY_EARS_DURABILITY_MAX);
+        BUNNY_EARS_EAT_BOOST_ENABLED = builder
+                .comment(BUNNY_EARS_EAT_BOOST_ENABLED_DESCRIPTION)
+                .define("bunnyEarsEatBoostEnabled", BUNNY_EARS_EAT_BOOST_ENABLED_DEFAULT);
+        BUNNY_EARS_JUMP_BOOST_ENABLED = builder
+                .comment(BUNNY_EARS_JUMP_BOOST_ENABLED_DESCRIPTION)
+                .define("bunnyEarsJumpBoostEnabled", BUNNY_EARS_JUMP_BOOST_ENABLED_DEFAULT);
         builder.pop();
 
         builder.push("ChoppingHat");
@@ -400,6 +408,16 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     @Override
     public int getBunnyEarsDurability() {
         return getInteger(BUNNY_EARS_DURABILITY, BUNNY_EARS_DURABILITY_DEFAULT);
+    }
+
+    @Override
+    public boolean isBunnyEarsEatBoostEnabled() {
+        return getBoolean(BUNNY_EARS_EAT_BOOST_ENABLED, BUNNY_EARS_EAT_BOOST_ENABLED_DEFAULT);
+    }
+
+    @Override
+    public boolean isBunnyEarsJumpBoostEnabled() {
+        return getBoolean(BUNNY_EARS_JUMP_BOOST_ENABLED, BUNNY_EARS_JUMP_BOOST_ENABLED_DEFAULT);
     }
 
     @Override